feat: Extract logic for OAuthButtons. Add starter code for account linking
