+++ app/components/auth/LoginForm.tsx
@@ -1,7 +1,7 @@
 // Source: https://github.com/mckaywrigley/chatbot-ui/blob/d60e1f3ee9d2caf8c9aab659791b841690183b2d/app/%5Blocale%5D/login/page.tsx#L145
 
 import React, { FormEvent, useState } from "react"
-import { redirect, useRouter } from "next/navigation"
+import { useRouter } from "next/navigation"
 import { supabase } from "@/app/lib/supabase/browser-client"
 import {
   AuthFormType,
@@ -11,8 +11,7 @@ import { PasswordInput } from "@/app/components/input"
 import { Routes } from "@/app/lib/constants"
 import { toast } from "sonner"
 import { Button, Checkbox, Divider, Input, Link } from "@nextui-org/react"
-import { Icon } from "@iconify-icon/react"
-import { Provider } from "@supabase/supabase-js"
+import OAuthButtons from "@/app/components/auth/OAuthButtons"
 
 export function LoginForm({
   setAuthFormType
@@ -24,26 +23,16 @@ export function LoginForm({
   const router = useRouter()
   const { closeAuthModal } = useAuthModal()
 
-  // TODO: Link anonymous accounts
-  async function handleOauthLogin(provider: Provider) {
-    const { data, error } = await supabase.auth.signInWithOAuth({
-      provider,
-      options: {}
-    })
-
-    if (data.url) {
-      redirect(data.url) // use the redirect API for your server framework
-    }
-  }
-
   async function handleLogin(event: FormEvent<HTMLFormElement>) {
     event.preventDefault()
     const formData = new FormData(event.currentTarget)
     const formJson = Object.fromEntries(formData.entries())
+    const email = formJson["email"] as string
+    const password = formJson["password"] as string
 
     const { data, error } = await supabase.auth.signInWithPassword({
-      email: formJson["email"] as string,
-      password: formJson["password"] as string
+      email: email,
+      password: password
     })
 
     // Show error message and return early if the login failed
@@ -64,6 +53,16 @@ export function LoginForm({
         homeWorkspaceError?.message || "An unexpected error occurred"
       )
     }
+
+    // if (isAnonymous) {
+    //   await supabase.auth.updateUser({ email: email })
+    //
+    //   // verify the user's email by clicking on the email change link
+    //   // or entering the 6-digit OTP sent to the email address
+    //
+    //   // Once the user has been verified, update the password
+    //   await supabase.auth.updateUser({ password: password })
+    // }
 
     // Handle successful login
     toast.success("Successfully logged in")
@@ -114,58 +113,7 @@ export function LoginForm({
         <p className="shrink-0 text-tiny text-default-500">OR</p>
         <Divider className="flex-1" />
       </div>
-      <div className="flex grow gap-2">
-        {/*<Button*/}
-        {/*  isIconOnly*/}
-        {/*  variant="bordered"*/}
-        {/*  onClick={() => handleOauthLogin("google")}*/}
-        {/*  className="grow"*/}
-        {/*>*/}
-        {/*  <Icon icon="logos:google-icon" className="text-xl" />*/}
-        {/*</Button>*/}
-        {/*<Button*/}
-        {/*  isIconOnly*/}
-        {/*  variant="bordered"*/}
-        {/*  className="grow"*/}
-        {/*  onClick={() => handleOauthLogin("apple")}*/}
-        {/*>*/}
-        {/*  <Icon className="text-xl dark:invert" icon="logos:apple" />*/}
-        {/*</Button>*/}
-        {/*<Button*/}
-        {/*  isIconOnly*/}
-        {/*  variant="bordered"*/}
-        {/*  className="grow"*/}
-        {/*  onClick={() => handleOauthLogin("discord")}*/}
-        {/*>*/}
-        {/*  <Icon className="text-xl" icon="logos:discord-icon" />*/}
-        {/*</Button>*/}
-        {/*<Button*/}
-        {/*  isIconOnly*/}
-        {/*  variant="bordered"*/}
-        {/*  className="grow"*/}
-        {/*  onClick={() => handleOauthLogin("github")}*/}
-        {/*>*/}
-        {/*  <Icon className="text-xl dark:invert" icon="logos:github-icon" />*/}
-        {/*</Button>*/}
-        <Button
-          variant="bordered"
-          className="grow"
-          onClick={() => handleOauthLogin("discord")}
-          startContent={<Icon className="text-xl" icon="logos:discord-icon" />}
-        >
-          Discord
-        </Button>
-        <Button
-          variant="bordered"
-          className="grow"
-          onClick={() => handleOauthLogin("github")}
-          startContent={
-            <Icon className="text-xl dark:invert" icon="logos:github-icon" />
-          }
-        >
-          GitHub
-        </Button>
-      </div>
+      <OAuthButtons />
       <p className="pt-2 text-center text-small">
         New to Metadachi?&nbsp;
         <Link
